Попытка разбить имеющийся код на модули (сборка через commonJS)
.В папке filter то, что касается фильтров, в папке pictures - картинок.
В utils - вспомогательные функции.
