+++ src/resizer.js
@@ -19,10 +19,6 @@
     this._frame.style.left = '50%';
     this._frame.style.top = '50%';
     this._frame.style.transform = 'translate(-50%, -50%)';
-    this._frame.style.position = 'absolute';
-    this._frame.style.left = '50%';
-    this._frame.style.top = '50%';
-    this._frame.style.transform = 'translate(-50%, -50%)';
 
     this._drawDotted = function(radius, freaquency, canvasElement) {
       canvasElement.setAttribute('width', this._resizeConstraint.side + 4 * radius + freaquency);
@@ -119,24 +115,6 @@
       // Очистка изображения.
       this._ctx.clearRect(0, 0, this._container.width, this._container.height);
 
-      // Параметры линии.
-      // NB! Такие параметры сохраняются на время всего процесса отрисовки
-      // canvas'a поэтому важно вовремя поменять их, если нужно начать отрисовку
-      // чего-либо с другой обводкой.
-
-      // Толщина линии.
-      //this._ctx.lineWidth = 6;
-
-      // Цвет обводки.
-      //this._ctx.strokeStyle = '#ffe753';
-
-      // Размер штрихов. Первый элемент массива задает длину штриха, второй
-      // расстояние между соседними штрихами.
-      //this._ctx.setLineDash([15, 10]);
-
-      // Смещение первого штриха от начала линии.
-      //this._ctx.lineDashOffset = 7;
-
       // Сохранение состояния канваса.
       // Подробней см. строку 132.
       this._ctx.save();
@@ -152,14 +130,6 @@
       // Координаты задаются от центра холста.
       this._ctx.drawImage(this._image, displX, displY);
 
-      // Отрисовка прямоугольника, обозначающего область изображения после
-      // кадрирования. Координаты задаются от центра.
-      /*this._ctx.strokeRect(
-          (-this._resizeConstraint.side / 2) - this._ctx.lineWidth / 2,
-          (-this._resizeConstraint.side / 2) - this._ctx.lineWidth / 2,
-          this._resizeConstraint.side - this._ctx.lineWidth / 2,
-          this._resizeConstraint.side - this._ctx.lineWidth / 2);*/
-
       // Восстановление состояния канваса, которое было до вызова ctx.save
       // и последующего изменения системы координат. Нужно для того, чтобы
       // следующий кадр рисовался с привычной системой координат, где точка
@@ -167,26 +137,6 @@
       // некорректно сработает даже очистка холста или нужно будет использовать
       // сложные рассчеты для координат прямоугольника, который нужно очистить.
       this._ctx.restore();
-
-      /*this._ctx.beginPath();
-      this._ctx.fillStyle = 'rgba(0, 0, 0, 0.8)';
-      this._ctx.moveTo(0, 0);
-      this._ctx.lineTo(this._container.width, 0);
-      this._ctx.lineTo(this._container.width, this._container.height);
-      this._ctx.lineTo(0, this._container.height);
-      this._ctx.lineTo(0, 0);
-      this._ctx.moveTo(this._container.width / 2 - this._resizeConstraint.side / 2 - this._ctx.lineWidth,
-          this._container.height / 2 - this._resizeConstraint.side / 2 - this._ctx.lineWidth);
-      this._ctx.lineTo(this._container.width / 2 + this._resizeConstraint.side / 2 - this._ctx.lineWidth / 2,
-          this._container.height / 2 - this._resizeConstraint.side / 2 - this._ctx.lineWidth);
-      this._ctx.lineTo(this._container.width / 2 + this._resizeConstraint.side / 2 - this._ctx.lineWidth / 2,
-          this._container.height / 2 + this._resizeConstraint.side / 2 - this._ctx.lineWidth / 2);
-      this._ctx.lineTo(this._container.width / 2 - this._resizeConstraint.side / 2 - this._ctx.lineWidth,
-          this._container.height / 2 + this._resizeConstraint.side / 2 - this._ctx.lineWidth / 2);
-      this._ctx.lineTo(this._container.width / 2 - this._resizeConstraint.side / 2 - this._ctx.lineWidth,
-          this._container.height / 2 - this._resizeConstraint.side / 2 - this._ctx.lineWidth);
-      this._ctx.closePath();
-      this._ctx.fill('evenodd');*/
       this._drawDotted(3, 15, this._frame);
       this._ctx.font = '20px Tahoma';
       this._ctx.fillStyle = 'white';
